moved feed and other related data models to shared module
