+++ shared/src/commonMain/kotlin/org/listenbrainz/shared/model/feed/FeedEventVisibilityData.kt
@@ -1,10 +1,10 @@
-package org.listenbrainz.android.model.feed
+package org.listenbrainz.shared.model.feed
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class FeedEventDeletionData (
+data class FeedEventVisibilityData (
     @SerialName("event_type") val eventType : String? = null,
-    @SerialName("id") val eventId   : String? = null
+    @SerialName("event_id") val eventId   : String? = null
 )
