+++ shared/src/commonMain/kotlin/org/listenbrainz/shared/model/feed/FeedPayload.kt
@@ -1,4 +1,4 @@
-package org.listenbrainz.android.model.feed
+package org.listenbrainz.shared.model.feed
 
 import androidx.compose.runtime.Immutable
 import kotlinx.serialization.SerialName
